+++ apps/forms/app/(api)/v1/session/[session]/field/[field]/file/upload/signed-url/route.ts
@@ -1,9 +1,11 @@
+import { client } from "@/lib/supabase/server";
 import { session_storage_createSignedUploadUrl } from "@/services/form/storage";
 import type {
   CreateSessionSignedUploadUrlRequest,
   FormsApiResponse,
   SessionSignedUploadUrlData,
 } from "@/types/private/api";
+import assert from "assert";
 import { NextRequest, NextResponse } from "next/server";
 
 export async function POST(
@@ -25,16 +27,36 @@ export async function POST(
   // TODO: validate if anonymous user is owner of this session
   // TODO: validate if session is open
 
-  const { data, error } = await session_storage_createSignedUploadUrl({
-    session_id: session_id,
-    field_id: field_id,
-    file: file,
-    config: {},
-  });
+  const { data, error } = await client
+    .from("response_session")
+    .select(
+      `id, form:form( fields:form_field( id, storage ), supabase_connection:connection_supabase(*) )`
+    )
+    .eq("id", session_id)
+    .single();
+
+  if (error || !data) {
+    throw error;
+  }
+
+  const { form } = data;
+  assert(form, "form not found");
+
+  const field = form.fields.find((field) => field.id === field_id);
+  assert(field, "form not found");
+
+  const { data: signeduploadurldata, error: signerr } =
+    await session_storage_createSignedUploadUrl({
+      session_id: session_id,
+      field: field,
+      connection: { supabase_connection: form.supabase_connection },
+      file: file,
+      config: {},
+    });
 
   return NextResponse.json(<FormsApiResponse<SessionSignedUploadUrlData>>{
-    data: data,
-    error: error,
+    data: signeduploadurldata,
+    error: signerr,
   });
 }
 
@@ -57,22 +79,42 @@ export async function PUT(
   // TODO: validate if anonymous user is owner of this session
   // TODO: validate if session is open
 
-  const { data, error } = await session_storage_createSignedUploadUrl({
-    session_id: session_id,
-    field_id: field_id,
-    file: file,
-    config: {
-      unique: true,
-    },
-  });
+  const { data, error } = await client
+    .from("response_session")
+    .select(
+      `id, form:form( fields:form_field( id, storage ), supabase_connection:connection_supabase(*) )`
+    )
+    .eq("id", session_id)
+    .single();
 
-  if (error) {
+  if (error || !data) {
+    throw error;
+  }
+
+  const { form } = data;
+  assert(form, "form not found");
+
+  const field = form.fields.find((field) => field.id === field_id);
+  assert(field, "form not found");
+
+  const { data: signeduploadurldata, error: signerr } =
+    await session_storage_createSignedUploadUrl({
+      session_id: session_id,
+      field: field,
+      connection: { supabase_connection: form.supabase_connection },
+      file: file,
+      config: {
+        unique: true,
+      },
+    });
+
+  if (signerr) {
     console.error("session/sign-upload-urls", error);
     return NextResponse.error();
   }
 
   return NextResponse.json(<FormsApiResponse<SessionSignedUploadUrlData>>{
-    data: data,
-    error: error,
+    data: signeduploadurldata,
+    error: signerr,
   });
 }
